Extract provider interface, update test
